+++ libft/ft_memset.c
@@ -6,15 +6,15 @@
 /*   By: sbouheni <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 08:45:57 by sbouheni          #+#    #+#             */
-/*   Updated: 2022/11/10 14:15:38 by sbouheni         ###   ########.fr       */
+/*   Updated: 2022/11/11 15:14:15 by sbouheni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *b, int c, int len)
+void	*ft_memset(void *b, int c, size_t len)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < len)
